+++ deduce/annotator.py
