+++ deduce/redact.py
